requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.6
python-dotenv==1.0.0

# Database
supabase==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
aiohttp==3.9.1
requests==2.31.0
pydantic==2.5.0
websockets==12.0

# CORS & Middleware
fastapi-cors==0.0.6

# Task Scheduler
apscheduler==3.10.4

# Additional utilities
Jinja2==3.1.2
